requirements.txt
Babel==2.7.0
Django==2.2
django-debug-toolbar==2.1
django-extensions==2.2.5
django-model-utils==4.0.0
django-phonenumber-field==4.0.0
phonenumberslite==8.11.1
python-stdnum==1.12
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
Werkzeug==0.16.0
